refactor: extract shared @lokascript/domain-config package
Adapted from feat/mcp-multilingual-intent commit 810b0de5. Original
commit shimmed both @hyperfixi/mcp-server and @lokascript/mcp-
multilingual-intent; the latter doesn't exist on main yet, so this
commit drops those hunks and consolidates only within mcp-server.

Single source of truth:
- @lokascript/domain-config (new) exports createDomainRegistry()
  (with scanConfig metadata) and DOMAIN_PRIORITY.
- mcp-server's domain-registry-setup.ts becomes a re-export shim
  (192 lines → 9 lines).
- mcp-server's dispatcher.ts imports DOMAIN_PRIORITY (was inline).

When mcp-multilingual-intent lands in a follow-up PR, it will import
from this shared package — no further changes needed here.

CI: adds Build domain-config step (after domain-toolkit) + artifact
upload path. check-ci-build-order: OK.

Verification:
- mcp-server: 415/415 tests pass (16 suites).
- domain-config typecheck + build clean.
